+++ src/compas_rhino/geometry/brep/brep.py
@@ -8,6 +8,7 @@ from compas_rhino.conversions import xform_to_rhino
 from compas_rhino.conversions import frame_to_rhino
 from compas_rhino.conversions import cylinder_to_rhino
 from compas_rhino.conversions import sphere_to_rhino
+from compas_rhino.conversions import mesh_to_rhino
 
 import Rhino
 
@@ -224,6 +225,23 @@ class RhinoBrep(Brep):
         """
         rhino_cylinder = cylinder_to_rhino(cylinder)
         return cls.from_native(rhino_cylinder.ToBrep(True, True))
+
+    @classmethod
+    def from_mesh(cls, mesh):
+        """Create a RhinoBrep from a mesh.
+
+        Parameters
+        ----------
+        mesh : :class:`~compas.datastructures.Mesh`
+            The source mesh.
+
+        Returns
+        -------
+        :class:`~compas_rhino.geometry.RhinoBrep`
+
+        """
+        rhino_mesh = mesh_to_rhino(mesh)
+        return cls.from_native(Rhino.Geometry.Brep.CreateFromMesh(rhino_mesh, True))
 
     # ==============================================================================
     # Methods